requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
billiard==4.2.1
boto3==1.37.19
botocore==1.37.19
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.1
cssselect2==0.8.0
cycler==0.12.1
distro==1.9.0
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.1.6
django-allauth==65.6.0
django-cors-headers==4.7.0
django-debug-toolbar==5.1.0
django-extensions==3.2.3
django-filter==25.1
django-sslserver==0.22
django-storages==1.14.5
django-taggit==6.1.0
django_csp==3.8
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
fonttools==4.56.0
graphviz==0.20.3
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
jmespath==1.0.1
joblib==1.5.1
kiwisolver==1.4.8
kombu==5.5.0
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mysql-connector-python==9.2.0
numpy==2.2.2
openai==1.69.0
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.2
pandas==2.2.2
patsy==1.0.1
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
py-enigma==0.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.11.0
pydantic_core==2.33.0
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
PyJWT==2.9.0
pyOpenSSL==25.0.0
pyparsing==3.2.1
pypdf==5.8.0
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
reportlab==4.4.2
requests==2.32.3
s3transfer==0.11.4
scikit-learn==1.7.0
scipy==1.15.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
statsmodels==0.14.4
svglib==1.5.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
whitenoise==6.9.0
xhtml2pdf==0.2.17
